Split tools registry and tighten types
