+++ src/journal/journal.ts
@@ -2,6 +2,7 @@ import fs from "node:fs/promises";
 import path from "node:path";
 import { appendJsonl } from "../util/jsonl.js";
 import { redact } from "../util/redaction.js";
+import { isErrnoException, isRecord } from "../util/types.js";
 
 export class SessionJournal {
   constructor(
@@ -28,13 +29,17 @@ export class TradeJournal {
     const filePath = path.join(this.baseDir, `${date}.jsonl`);
     try {
       const raw = await fs.readFile(filePath, "utf8");
-      return raw
-        .split("\n")
-        .map((line) => line.trim())
-        .filter(Boolean)
-        .map((line) => JSON.parse(line) as Record<string, unknown>);
+      const entries: Record<string, unknown>[] = [];
+      for (const line of raw.split("\n").map((item) => item.trim())) {
+        if (!line) continue;
+        const parsed = JSON.parse(line);
+        if (isRecord(parsed)) {
+          entries.push(parsed);
+        }
+      }
+      return entries;
     } catch (err) {
-      if ((err as NodeJS.ErrnoException).code === "ENOENT") {
+      if (isErrnoException(err) && err.code === "ENOENT") {
         return [];
       }
       throw err;
