+++ src/util/redaction.ts
@@ -2,26 +2,20 @@ const SECRET_KEY_PATTERN =
   /(privateKey|secret|apiKey|token|authorization|auth|bearer)/i;
 const BYTE_LIKE = /(Uint8Array|Buffer)/;
 
-export type Redactable =
-  | Record<string, unknown>
-  | unknown[]
-  | string
-  | number
-  | boolean
-  | null;
+export type Redactable = unknown;
 
 function shouldRedactKey(key: string): boolean {
   return SECRET_KEY_PATTERN.test(key);
 }
 
-export function redact<T extends Redactable>(value: T, depth = 0): T {
+export function redact(value: Redactable, depth = 0): Redactable {
   if (depth > 8) return value;
   if (value === null || value === undefined) return value;
-  if (typeof value === "string") return value as T;
+  if (typeof value === "string") return value;
   if (typeof value !== "object") return value;
 
   if (Array.isArray(value)) {
-    return value.map((item) => redact(item as Redactable, depth + 1)) as T;
+    return value.map((item) => redact(item, depth + 1));
   }
 
   const obj = value as Record<string, unknown>;
@@ -38,7 +32,7 @@ export function redact<T extends Redactable>(value: T, depth = 0): T {
         continue;
       }
     }
-    out[key] = redact(val as Redactable, depth + 1);
+    out[key] = redact(val, depth + 1);
   }
-  return out as T;
+  return out;
 }
