+++ src/util/types.ts
@@ -1,0 +1,9 @@
+export function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+
+export function isErrnoException(
+  value: unknown,
+): value is NodeJS.ErrnoException {
+  return value instanceof Error && "code" in value;
+}
